backend/requirements.txt
Flask
flask-cors
requests
PyMySQL
python-dotenv
cloudinary
google-generativeai
Pillow
